+++ src/features/export/constants/index.ts
@@ -1,0 +1,1 @@
+export { VIDEO_STAGE_LABELS } from './stage-labels'
